refactor(ui): colocate menu factories with schema
