+++ packages/ui/src/menu/history.menu.ts
@@ -15,8 +15,8 @@
  */
 
 import type { IAccessor } from '@univerjs/core';
-import type { LocaleKey } from '../../locale/types';
-import type { IMenuButtonItem } from '../../services/menu/menu';
+import type { LocaleKey } from '../locale/types';
+import type { IMenuButtonItem } from '../services/menu/menu';
 import {
     EDITOR_ACTIVATED,
     FOCUSING_FX_BAR_EDITOR,
@@ -28,7 +28,7 @@ import {
 } from '@univerjs/core';
 import { combineLatest, merge, of } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { MenuItemType } from '../../services/menu/menu';
+import { MenuItemType } from '../services/menu/menu';
 
 const undoRedoDisableFactory$ = (accessor: IAccessor, isUndo: boolean) => {
     const undoRedoService = accessor.get(IUndoRedoService);
