+++ packages/ui/src/menu/schema.ts
@@ -18,10 +18,10 @@ import type { MenuSchemaType } from '../services/menu/menu-manager.service';
 import { RedoCommand, UndoCommand } from '@univerjs/core';
 import { OpenFeatureSearchOperation } from '../commands/operations/open-feature-search.operation';
 import { ToggleShortcutPanelOperation } from '../commands/operations/toggle-shortcut-panel.operation';
-import { FeatureSearchMenuItemFactory } from '../controllers/feature-search/menu';
-import { RedoMenuItemFactory, UndoMenuItemFactory } from '../controllers/menus/menus';
-import { ShortcutPanelMenuItemFactory } from '../controllers/shortcut-display/menu';
 import { RibbonStartGroup } from '../services/menu/types';
+import { FeatureSearchMenuItemFactory } from './feature-search.menu';
+import { RedoMenuItemFactory, UndoMenuItemFactory } from './history.menu';
+import { ShortcutPanelMenuItemFactory } from './shortcut-panel.menu';
 
 export const menuSchema: MenuSchemaType = {
     [RibbonStartGroup.HISTORY]: {
